+++ rest_framework_json_api/metadata.py
@@ -7,7 +7,7 @@ from rest_framework.metadata import SimpleMetadata
 from rest_framework.settings import api_settings
 from rest_framework.utils.field_mapping import ClassLookupDict
 
-from rest_framework_json_api.utils import format_value, get_related_resource_type
+from rest_framework_json_api.utils import format_field_name, get_related_resource_type
 
 
 class JSONAPIMetadata(SimpleMetadata):
@@ -93,7 +93,7 @@ class JSONAPIMetadata(SimpleMetadata):
 
         return OrderedDict(
             [
-                (format_value(field_name), self.get_field_info(field))
+                (format_field_name(field_name), self.get_field_info(field))
                 for field_name, field in serializer.fields.items()
             ]
         )