+++ rest_framework_json_api/views.py
@@ -25,9 +25,9 @@ from rest_framework_json_api.serializers import ResourceIdentifierObjectSerializ
 from rest_framework_json_api.utils import (
     Hyperlink,
     OrderedDict,
-    format_value,
     get_included_resources,
     get_resource_type_from_instance,
+    undo_format_link_segment,
 )
 
 
@@ -187,7 +187,7 @@ class RelatedMixin(object):
 
     def get_related_field_name(self):
         field_name = self.kwargs["related_field"]
-        return format_value(field_name, "underscore")
+        return undo_format_link_segment(field_name)
 
     def get_related_instance(self):
         parent_obj = self.get_object()