+++ CHANGELOG.md
@@ -14,6 +14,15 @@ any parts of the framework not mentioned in the documentation should generally b
 
 * Added support for Django 3.2.
 
+### Fixed
+
+* Preserve field names when no formatting is configured.
+
+### Deprecated
+
+* Deprecated default `format_type` argument of `rest_framework_json_api.utils.format_value`. Use `rest_framework_json_api.utils.format_field_name` or specify specifc `format_type` instead.
+* Deprecated `format_type` argument of `rest_framework_json_api.utils.format_link_segment`. Use `format_value` instead.
+
 ## [4.1.0] - 2021-03-08
 
 ### Added